Enhance model readiness checks in BackgroundTask and ModelPoolField to prevent crashes on uninitialized models

+++ apps/isocron/__init__.py
@@ -7,7 +7,9 @@ from datetime import time, date, datetime, timedelta
 
 import isodate
 from django.utils import timezone
-
+from django.apps import apps
+from django.db.utils import ProgrammingError
+from django.core.exceptions import ImproperlyConfigured, AppRegistryNotReady
 from misc.utils import load
 
 ISO_PARSER = {
@@ -120,29 +122,30 @@ def autodiscover():
 
     """
     try:
-        from .models import BackgroundTask
-        load('cron')
-        tasks = BaseCronJob.get_all()
-        existing = set(BackgroundTask.objects.values_list('name', flat=True))
-        new_tasks = set(tasks.keys()) - existing
+        if apps.ready:
+            from .models import BackgroundTask
+            load('cron')
+            tasks = BaseCronJob.get_all()
+            existing = set(BackgroundTask.objects.values_list('name', flat=True))
+            new_tasks = set(tasks.keys()) - existing
 
-        to_create = [
-            BackgroundTask(
-                name=name, run_every=tasks[name].run_every, run_at=tasks[name].run_at,
-                retry_after=tasks[name].retry_after, keep_logs=tasks[name].keep_logs,
-                description=(tasks[name].__doc__ or "").replace('\n', ' ').strip()
-            )
-            for name in new_tasks
-        ]
+            to_create = [
+                BackgroundTask(
+                    name=name, run_every=tasks[name].run_every, run_at=tasks[name].run_at,
+                    retry_after=tasks[name].retry_after, keep_logs=tasks[name].keep_logs,
+                    description=(tasks[name].__doc__ or "").replace('\n', ' ').strip()
+                )
+                for name in new_tasks
+            ]
 
-        if to_create:
-            BackgroundTask.objects.bulk_create(to_create)
+            if to_create:
+                BackgroundTask.objects.bulk_create(to_create)
 
-        to_remove = existing - set(tasks.keys())
-        if to_remove:
-            BackgroundTask.objects.filter(name__in=list(to_remove)).delete()
+            to_remove = existing - set(tasks.keys())
+            if to_remove:
+                BackgroundTask.objects.filter(name__in=list(to_remove)).delete()
 
-    except Exception:
+    except (AppRegistryNotReady, ProgrammingError) as e:
         # If there is an error, we do not want to crash the server, just log it
         import logging
         logger = logging.getLogger(__name__)
